requirements.txt
streamlit==1.26.0
PyPDF2==3.0.1
docx2txt==0.8
numpy==1.25.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.33.0
torch==2.0.1
